+++ src/main/java/ca/uvic/seng330/assn3/model/devices/SmartPlug.java
@@ -32,8 +32,15 @@ public class SmartPlug extends Device {
       setStatus(Status.OFF);
       Logging.logWithID("plug turned off", getIdentifier(), Level.INFO);
     } else {
-      Logging.logWithID("Toggling plug failed. Not in ON or OFF state. Likely ERROR state.", getIdentifier(), Level.ERROR);
+      Logging.logWithID(
+          "Toggling plug failed. Not in ON or OFF state. Likely ERROR state.",
+          getIdentifier(),
+          Level.ERROR);
     }
+    this.getHub()
+        .notification(
+            "SmartPlug " + this.getLabel() + " has been turned " + this.getStatus().toString(),
+            this.getIdentifier());
   }
 
   @Override